paypro: fix browser signatures with KJUR. move pem/der functions to common.
